Test Level Editing
Finished More editing on the Test Level and
fixed a bug on the test level that stopped third person character from spawning.

Note: the large metal wall I have added at the end of the level is intended to be climbable by the player when adding a magnetic grip ability.